Updated environment and separated evaluator class
